// src/apps/travellrs/enums/travellrs.enum.ts
export enum TRAVELLRS_EVENT_TYPES {
  NEW_MARKER = 'NEW_MARKER',

  NEW_MARKER_REACTION = 'NEW_MARKER_REACTION',
  NEW_MARKER_UNREACTION = 'NEW_MARKER_UNREACTION',
  NEW_MARKER_CHANGED_REACTION = 'NEW_MARKER_CHANGED_REACTION',

  NEW_MARKER_COMMENT = 'NEW_MARKER_COMMENT',
  NEW_MARKER_COMMENT_REACTION = 'NEW_MARKER_COMMENT_REACTION',
  NEW_MARKER_COMMENT_UNREACTION = 'NEW_MARKER_COMMENT_UNREACTION',
  NEW_MARKER_COMMENT_CHANGED_REACTION = 'NEW_MARKER_COMMENT_CHANGED_REACTION',

  NEW_MARKER_COMMENT_REPLY = 'NEW_MARKER_COMMENT_REPLY',
  NEW_MARKER_COMMENT_REPLY_REACTION = 'NEW_MARKER_COMMENT_REPLY_REACTION',
  NEW_MARKER_COMMENT_REPLY_UNREACTION = 'NEW_MARKER_COMMENT_REPLY_UNREACTION',
  NEW_MARKER_COMMENT_REPLY_CHANGED_REACTION = 'NEW_MARKER_COMMENT_REPLY_CHANGED_REACTION',
}

export enum TRAVELLRS_NOTIFICATION_TARGET_TYPES {
  MARKER = 'MARKER',
  MARKER_COMMENT = 'MARKER_COMMENT',
  MARKER_COMMENT_REPLY = 'MARKER_COMMENT_REPLY',
}

export enum TRAVELLRS_SUBSCRIPTION_TARGET_ACTIONS {
  
}

export enum TRAVELLRS_SUBSCRIPTION_TARGET_ACTIONS_INFO {
  
}

export enum MarkerVisibility {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
}